Creation d'un "Cooperative Visitor" sur le scene de graph pour pouvoir ajouter de nouvelles fonctionnalites aux objets sans changer leurs structures. /!\ la GUI utilise maintenant des visiteurs pour se positionner, test le focus etc... mais l'implementation bug, il faut debugger ca !
+ Portage windows, pour le moment on ne compile qu'avec des lib static et le Tuto1 segv lorsqu'on quitte l'application.

+++ Nc/Graphics/Camera/StandardCamera3d.cpp
@@ -96,7 +96,7 @@ const char *StandardCamera3d::XpmHandClose[] =
     "0,0"
 };
 
-StandardCamera3d::StandardCamera3d(Graphic::Window *win, double ratioAspect, double nearD, double farD, double fielOfView, Pattern p)
+StandardCamera3d::StandardCamera3d(Graphic::Window *win, float ratioAspect, float nearD, float farD, float fielOfView, Pattern p)
     : Camera3d(ClassName(), ratioAspect, nearD, farD, fielOfView),
       _mouveButton(System::Mouse::Right), _pattern(p),
       _inhibitMovement(false), _drawFrustum(false),
@@ -120,12 +120,12 @@ StandardCamera3d::StandardCamera3d(Graphic::Window *win, double ratioAspect, dou
 }
 
 StandardCamera3d::StandardCamera3d(Graphic::Window *win, Pattern p)
-    : Camera3d(ClassName(), (float)win->Width()/(float)win->Height(), 0.1, 1000, 70),
+    : Camera3d(ClassName(), (float)win->Width()/(float)win->Height(), 0.1f, 1000.f, 70.f),
       _mouveButton(System::Mouse::Right), _pattern(p),
       _inhibitMovement(false), _drawFrustum(false),
       _angles(-90, -15)
 {
-    // create cursors
+	// create cursors
     _cursorOpen = win->NewCursor();
     _cursorOpen->LoadFromXpm(XpmHandOpen);
     _cursorClose = win->NewCursor();
